renamed java parser dir, remove extra deps
