+++ package.json
@@ -29,7 +29,6 @@
     "body-parser": "1.13.3",
     "bootstrap": "^3.3.5",
     "bootswatch": "^3.3.5",
-    "bytebuffer": "^4.0.0",
     "c3": "^0.4.10",
     "cal-heatmap": "^3.5.2",
     "compression": "^1.5.2",
@@ -51,7 +50,6 @@
     "jade": "~1.11.0",
     "jquery": "^2.1.4",
     "kue": "~0.9.4",
-    "long": "^2.2.5",
     "moment": "^2.10.6",
     "monk": "1.0.1",
     "ndjson": "^1.4.1",
@@ -62,18 +60,12 @@
     "paypal-rest-sdk": "~1.6",
     "pm2": "^0.14.6",
     "poet": "~2.0.1",
-    "protobufjs": "^4.0.0",
     "qTip2": "yasp-dota/qtip2",
     "redis": "0.12.1",
     "request": "~2.60.0",
     "request-progress": "^0.3.1",
-    "select2": "^4.0.0",
-    "select2-bootstrap-theme": "select2/select2-bootstrap-theme",
     "selectize": "^0.12.1",
     "sentiment": "^1.0.2",
-    "snappy": "^3.2.0",
-    "socket.io": "^1.3.6",
-    "socket.io-client": "^1.3.6",
     "steam": "0.6.8",
     "winston": "1.0.1",
     "wordcloud": "1.0.1"
@@ -85,7 +77,6 @@
     "file-loader": "^0.8.4",
     "istanbul": "~0.3.17",
     "json-loader": "^0.5.2",
-    "look": "^0.1.3",
     "mocha": "~2.2.5",
     "mongodb": "~2.0.40",
     "nock": "~2.10.0",
