--- parser/pom.xml
@@ -1,90 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>co.yasp</groupId>
-    <artifactId>stats</artifactId>
-    <version>0.1.0</version>
-    <packaging>jar</packaging>
-    <name>stats</name>
-    <developers>
-        <developer>
-            <id>albertcui</id>
-            <name>Albert Cui</name>
-            <email>albertqcui@gmail.com</email>
-        </developer>
-        <developer>
-            <id>howardc93</id>
-            <name>Howard Chung</name>
-            <email>howardc93@gmail.com</email>
-        </developer>
-    </developers>
-    <properties>
-        <maven.compiler.target>1.7</maven.compiler.target>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-    <repositories>
-        <repository>
-            <id>jitpack.io</id>
-            <url>https://jitpack.io</url>
-        </repository>
-        <repository>
-            <id>sonatype.oss.snapshots</id>
-            <name>Sonatype OSS Snapshot Repository</name>
-            <url>http://oss.sonatype.org/content/repositories/snapshots</url>
-            <releases>
-                <enabled>false</enabled>
-            </releases>
-            <snapshots>
-                <enabled>true</enabled>
-            </snapshots>
-        </repository>
-    </repositories>
-    <dependencies>
-        <dependency>
-            <groupId>com.github.skadistats</groupId>
-            <artifactId>clarity</artifactId>
-            <version>-SNAPSHOT</version>
-        </dependency>
-        <!--
-        <dependency>
-            <groupId>com.github.skadistats</groupId>
-            <artifactId>clarity</artifactId>
-            <version>70956f5a89fca56f3a292e739ea3eda96a33fd8c</version>
-        </dependency>
-        <dependency>
-            <groupId>com.skadistats</groupId>
-            <artifactId>clarity</artifactId>
-            <version>2.1-SNAPSHOT</version>
-        </dependency>
-        -->
-        <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.3.1</version>
-        </dependency>
-    </dependencies>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>2.2</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                        <configuration>
-                            <transformers>
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>yasp.Main</mainClass>
-                                </transformer>
-                            </transformers>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
